+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added the classes INScrollView and INTableView with the ability to cancel touches on certain or all views when dragging
 - Fixes monthsBetweenDate: in NSDate+INExtensions
 - Added the possibility to instantiate an INCoreDataManager instance with a specific model version
+- Fixed the duplicateStoreToUrl: method in INCoreDataManager
 - ...